pyproject.toml
[project]
name = "withthanks"
version = "0.1.0"
description = "Personalised donor thank-you video engine"
requires-python = ">=3.12"
dependencies = [
    "Django>=6.0.2",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-yasg>=1.21.15",
    "elevenlabs>=2.37.0",
    "ffmpeg-python>=0.2.0",
    "resend>=2.23.0",
    "python-dotenv>=1.2.2",
    "gunicorn>=25.1.0",
    "celery>=5.6.2",
    "redis>=7.2.1",
    "psycopg[binary]>=3.3.3",
    "pydantic>=2.12.5",
    "httpx>=0.28.1",
    "whitenoise>=6.12.0",
    "xhtml2pdf>=0.2.17",
    "openpyxl>=3.1.5",
    "django-tailwind-cli>=4.5.1",
]

[project.optional-dependencies]
# Cloudflare R2 storage (S3-compatible API — only needed if R2 is enabled)
r2 = [
    "boto3>=1.42.59",
    "django-storages[s3]>=1.14.6",
]

# ---------------------------------------------------------------------------
# uv
# ---------------------------------------------------------------------------
[tool.uv]

[dependency-groups]
dev = [
    "django-stubs[compatible-mypy]>=5.1",
    "djangorestframework-stubs>=3.15",
    "pyright>=1.1.408",
    "ruff>=0.15.4",
]

# ---------------------------------------------------------------------------
# ruff
# ---------------------------------------------------------------------------
[tool.ruff]
target-version = "py312"
line-length = 100
src = ["."]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
    "RUF", # ruff-specific rules
]
ignore = [
    "E501",   # line too long — ruff format handles it
    "B008",   # do not perform function calls in default arguments (Django ORM common pattern)
    "RUF012", # mutable class vars — common in Django models
]

[tool.ruff.lint.isort]
known-first-party = ["charity", "withthanks"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

# ---------------------------------------------------------------------------
# pyright
# ---------------------------------------------------------------------------
[tool.pyright]
pythonVersion = "3.12"
pythonPlatform = "Linux"
include = ["charity", "withthanks"]
exclude = ["**/__pycache__", "**/migrations", ".venv"]
venvPath = "."
venv = ".venv"
typeCheckingMode = "basic"

# Suppress noise from missing / incomplete stubs
reportMissingModuleSource = false
reportMissingImports = false
reportUnknownMemberType = false
reportUnknownVariableType = false
reportUnknownArgumentType = false
reportUnknownParameterType = false
reportPrivateImportUsage = false

# Known Django ORM false-positives without full stub coverage
reportAttributeAccessIssue = false
reportAssignmentType = false
reportIndexIssue = false
reportOptionalSubscript = false
reportCallIssue = false
reportArgumentType = false
reportReturnType = false
